Integrated existing vclamp with real time socket.
I have tried to change as little about the vclamp code as possible and maintain normal behaviour in non-RT mode. Perhaps the most incisive change is in the GNUMakefile, which now contains a modified version of the previously included GeNN makefile as well as a switch to build RT objects if requested.