Added support for user-defined models
This update divorces the static simulation/GA code from the user-defined and GeNN-generated model code. In the interest of a directory structure that reflects this, and in anticipation of streamlining in the sim/GA code, I have duplicated the sim/GA code from :/vclamp/model into :/simulation, thus freeing it from compatibility constraints with regards to fully simulated use.