Wrapped the realtime comedi function to prevent conflicts
On the realtime side, source files (but not headers) should #include "RC_rtai_comedi.h" instead of <rtai_comedi.h>. This prefixes the realtime comedi functions with RC_, allowing for a clear separation between realtime and non-realtime comedi functions.
On the non-realtime side, libcomedi is now dynamically linked in, so there is no more need to rely on dlopen()/dlsym() to get access to things like board name, subdevices etc.
softrtdaq.c will be altered accordingly in a future update.