Moved control over RunData into ExpLib, decoupled Exp from Lib
Constructing the ExperimentLibrary outside of Experiment objects and injecting
it as a dependency instead lets me work with a single library on various
kinds of experiment, rather than building a new library for each (and having
to deal with cleanup / clogging the GPU in the process). Case in point:
I can now extend error profiling into a separate tool, and eventually
build a whole set of similar tools that are independent except in their
dependence on an experiment library.