Added a third library for rapid sampled profiling.
The idea of this beast is to use randomly parametrised models both as
probes and as references, providing a large number of gradient evaluations
over a wide range of parameter sets. In aggregate, the values returned
give me an assessment similar to the smooth/steep intuition, only in a more
principled manner (smoothness -> accuracy of gradients; steepness -> median
gradient) and over a wider range of parameter sets.