[BREAKING] Fixed WavegenSelection ordering issue
This is a not-quite-fix for an incompatibility introduced in commit
5b8e8d, where a faulty WaveSource::stimulations() subroutine was
replaced with a correct one that revealed the bug fixed here.
Note that interoperability with pre-5b8e8d sessions is still compromised,
as that was off-by-one.
Pre-5b8e8d sessions are internally consistent, as far as I can tell.