Added Ornstein-Uhlenbeck noise to Simulator
This implementation is very similar to that in the noise conductances in
StdpC, only without the dependence on the membrane voltage or a reversal
potential. Since the aim is to inject some noise rather than to accurately
simulate a particular process, this seems acceptable. In experimental
conditions, noise in the clamp circuit probably arises from many sources,
both cellular (stochastic conductances, coupling, synaptic inputs) and
instrumental (electrode, transmission and amplifier). Their sum is likely
to be at least somewhat similar to a white noise (low tau) current input
as implemented.