Extend UniversalLibrary to support unique stim/obs/rund/target vars
Having these variables in __constant__ memory should make loading them up
a lot quicker, but more importantly, I don't need to push 10k identical
ones over when it comes to that.
The target trace itself is still just in normal device memory, that might
be worth moving to __constant__ as well - problem is, the constant cache
is not large enough for even a single trace (64kB max).