Improved SamplingProfiler processing kernel
I've tested the indexing thoroughly. The previous version
mis-indexed on writes by squishing the diagonal in a wrong way.
This wouldn't have affected the results much, but it's good to
get it right.

Performance gain is minimal (Kv21f, 2017.10.06, Selection 1/EK,
full range, interval 20, dt .25/20), which is surprising given
that there are nSamplesUsed*NPAIRS^2 fewer reads...