Added a deltabar calculation kernel to normalise contributions
As in the previous Wavegen implementation, using deltabar feels more
natural than adjusting the sigmas to potentially unsound levels.
The aim, of course, is to bring the deviation caused by a detuning into
a similar range for all parameters, so as to prevent certain parameters
from having too little or too much influence on the clustering.
Note, this is equivalent to the so-far unused 'norm' argument included in
clustering functions.